Improved: Truncate encoded string (OFBIZ-13167) (#844)
* Amélioration de la function de truncate string

On model form, when you use a field description with a text size like :

<form...>
    <field name="myfield" ...><display size="15"/></field>
</form>

and your field myField contains a string with a special character like "my char é, so bad truncate", the widget rendering displays "my char &ea...ate" because when truncate is executed, the string is encoded and some chars are encoded : "my char &eacut;, so bad truncate".

The size between the encoded string and the displayed string are different. So to fix this, we implemented a new method to truncate the content. This method identify each special char as one char, and truncate.

"my char é, so bad truncate" encoded to "my char &eacut;, so bad truncate" become "my char &eacut;, ...ate" and not "my char &eac...ate"

Thanks: Charles Steltzlen who help to solve it