Fixed: Remove error after test rollback (OFBIZ-13172)
When you run entity test with `.gradlew 'ofbiz -t component=entity' after the test, entity engine rollback all entity operation. During this process, an error appears due to immutable state on GenericValue rollbacked.

To solve it, we'll check first if a gv has been set as immutable. In this case we clone it before call the delegator.