Improved: Optimization on removeJob service (OFBIZ-13182)
On removeJob service when we'll remove a job sandbox, we also delete other element linked (RecurrenceInfo and RuntimeData).

Before their remove we control that no other job are connected to them, but currently we resolve the list and ensure that it empty.

To be fast, we just ask the database through a count and ensure that is zero.