Completed: Improve Helveticus theme (OFBIZ-12276)
This sets HELVETICUS_EMERALD as default theme. It follows the discussion at
https://lists.apache.org/thread/47bg0tr4w15dj3o55gzrsb4chkj227hl

I must say that I picked green (emerald) instead of blue (sapphire) used before
because it's my favorite colour... as far as I can remember... (that's how
favorite color works, right?)

Note that we could say that view from space the earth is blue and that's what
we (humanity) have most in common. But wait, I could say I prefer green because
it's the origin of life. Without plants nothing would live on this earth. And
it believe I also I like green, not not only because it's obviously life but
it also reassures me anytime.

Of course that's my choice and we can discuss it. Did we for blue in the past?