Improved: Allow to use GroovyDsl in FlexibleStringExpander (OFBIZ-13133) (#839)
Second improvement on this functionality with increase the security by analyse each script to control the presence of potential code injection.

The regexp to control is a property: security.deniedScriptletsTokens.
If a script match the regexp, OFBiz raise in log an alert with the script and the script hash. The script is disabled and can't run.

If you have a safe script who is matched by the regexp, you can add the hash given by OFBiz on the property: security.allowedScriptletHashes

We added a new test that scan all xml file to analyse groovy scriplet and return all unsafe scriptlet found. The test will fail if unsafe scriptlets was found.

Thanks to Gil Portenseigne and Jacques Le Roux for help and review